:art: split up to mirror package structure
- scipy.tests function are tested in scipy?
